In Class updates to Smash app.
Smash

+++ Smash/BrickView.swift
@@ -26,7 +26,7 @@ class BrickView: UIView {
         
         var context = UIGraphicsGetCurrentContext()
         
-        UIColor(white: 0.0, alpha: 0.3 * health + 0.1).set()
+        UIColor(white: 0.0, alpha: 0.3 * CGFloat(health) + 0.1).set()
         
         CGContextFillRect(context, rect)
